requirements.txt
tensorflow==2.2.0
bert4keras==0.10.5
keras==2.3.1
jieba
tqdm
numpy
textrank4zh
xlrd2
xlwt
annoy==1.17.0